add authenticate

+++ student/views_with_api_class.py
@@ -1,45 +1,58 @@
-from .models import Student
 from rest_framework.views import APIView
-from rest_framework import status
+from .models import Student
 from rest_framework.response import Response
+from rest_framework import status
 from .serializers import StudentSerializer
 from django.shortcuts import get_object_or_404
 
 
-class StudentAPI(APIView):
-    def get(self, request, pk=None):
-        if pk is not None:
-            student = Student.objects.get(id=pk)
-            serializer = StudentSerializer(student)
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        students = Student.objects.all()
-        serializer = StudentSerializer(students, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+class StudentAPIView(APIView):
+    serializer_class = StudentSerializer
+    query_set = Student.objects.all()
+
+    def get_object(id):
+        pass
+
+    def get(self, request, id=None):
+        if id:
+            student = get_object_or_404(Student, id=id)
+            student_serializer = self.serializer_class(student)
+        else:
+
+            student_serializer = self.serializer_class(self.query_set, many=True)
+        return Response(student_serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        serializer = StudentSerializer(data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data,
-                            status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        student_serializer = self.serializer_class(data=request.data)
+        student_serializer.is_valid(raise_exception=True)
+        student_serializer.save()
+        return Response(student_serializer.data, status=status.HTTP_201_CREATED)
 
-    def put(self, request, pk):
-        student = Student.objects.get(pk=pk)
-        serializer = StudentSerializer(student, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+    def patch(self, request, id):
+        student = Student.objects.filter(id=id).first()
+        if student is None:
+            return Response({"details": "student is not found"},
+                            status=status.HTTP_404_NOT_FOUND)
+        student_serializer = self.serializer_class(student, data=request.data,
+                                               partial=True)
+        student_serializer.is_valid(raise_exception=True)
+        student_serializer.save()
+        return Response(student_serializer.data, status=status.HTTP_200_OK)
 
-    def patch(self, request, pk):
-        stu = Student.objects.get(pk=pk)
-        serializer = StudentSerializer(stu, data=request.data, partial=True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+    def put(self, request, id):
+        student = Student.objects.filter(id=id).first()
+        if student is None:
+            return Response({"details": "student is not found"},
+                            status=status.HTTP_404_NOT_FOUND)
+        student_serializer = self.serializer_class(student, data=request.data)
+        student_serializer.is_valid(raise_exception=True)
+        student_serializer.save()
+        return Response(student_serializer.data, status=status.HTTP_200_OK)
 
-    def delete(self, request, pk):
-        student = get_object_or_404(Student, pk=pk)
+    def delete(self, request, id):
+        student = Student.objects.filter(id=id).first()
+        if student is None:
+            return Response({"details": "student is not found"},
+                            status=status.HTTP_404_NOT_FOUND)
         student.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
